services/stac-api/requirements.txt
stac-fastapi.pgstac==6.2.2
fastapi>=0.119.0
uvicorn>=0.30.1
psycopg[binary]
psycopg-pool
